refactor the bowtie2 sub-module
