+++ q2_assembly/bowtie2/tests/test_indexing.py
@@ -14,47 +14,17 @@ from q2_types_genomics.per_sample_data import (ContigSequencesDirFmt,
                                                MultiMAGSequencesDirFmt)
 from qiime2.plugin.testing import TestPluginBase
 
-from q2_assembly.bowtie2.bowtie2 import (_get_subdir_from_path, _index_seqs,
-                                         index_contigs, index_mags,
-                                         _process_bowtie2_arg)
+from q2_assembly.bowtie2.indexing import (
+    _index_seqs, index_contigs, index_mags
+)
 
 
-class TestBowtie2(TestPluginBase):
+class TestBowtie2Indexing(TestPluginBase):
     package = 'q2_assembly.bowtie2.tests'
 
     def setUp(self):
         super().setUp()
         self.test_params_list = ['--large-index', '--bmax', '11']
-
-    def test_get_subdir_from_path_contig(self):
-        obs = _get_subdir_from_path('/path/to/dir/sample1_contigs.fa')
-        exp = 'sample1'
-        self.assertEqual(obs, exp)
-
-    def test_get_subdir_from_path_mag(self):
-        obs = _get_subdir_from_path('/path/to/dir/sample1/mag1.fa', 'mags')
-        exp = 'sample1/mag1'
-        self.assertEqual(obs, exp)
-
-    def test_get_subdir_from_path_unknown_type(self):
-        with self.assertRaisesRegex(
-                NotImplementedError, r'"unicorn" is not supported'):
-            _ = _get_subdir_from_path('/path/to/dir/mag1.fa', 'unicorn')
-
-    def test_process_bowtie2_arg_simple1(self):
-        obs = _process_bowtie2_arg('not_k_list', 123)
-        exp = ['--not-k-list', '123']
-        self.assertListEqual(obs, exp)
-
-    def test_process_bowtie2_arg_simple2(self):
-        with self.assertRaisesRegex(
-                Exception, r'.*type "\<class \'list\'\>" is not supported\.'):
-            _process_bowtie2_arg('k_list', [1, 2, 3])
-
-    def test_process_bowtie2_arg_bool(self):
-        obs = _process_bowtie2_arg('k_bool', True)
-        exp = ['--k-bool']
-        self.assertListEqual(obs, exp)
 
     @patch('subprocess.run')
     @patch('os.makedirs')
@@ -106,7 +76,7 @@ class TestBowtie2(TestPluginBase):
                 input_type='mags'
             )
 
-    @patch('q2_assembly.bowtie2._index_seqs')
+    @patch('q2_assembly.bowtie2.indexing._index_seqs')
     def test_index_contigs(self, p):
         input_contigs = ContigSequencesDirFmt(
             self.get_data_path('contigs'), 'r')
@@ -117,7 +87,7 @@ class TestBowtie2(TestPluginBase):
         p.assert_called_with(
             exp_contigs, ANY, self.test_params_list, 'contigs')
 
-    @patch('q2_assembly.bowtie2._index_seqs')
+    @patch('q2_assembly.bowtie2.indexing._index_seqs')
     def test_index_mags(self, p):
         input_mags = MultiMAGSequencesDirFmt(
             self.get_data_path('mags'), 'r')
